==> 4.Basketball-reference EFF and GmSc Calculation.ipynb <==
#Basketball-reference EFF and GmSc Calculation

# special IPython command to prepare the notebook for matplotlib
from collections import defaultdict

import pandas as pd
import matplotlib.pyplot as plt
import requests
import re
import string
from pattern import web

### Read Player Seaaon CSV

df_Season_Total = pd.read_csv('e:\\nba\\Season_Totals.csv', header=0)

df_Season_Total.info()
#df_Season_Total.describe()

### Read Season_Advance_Totals CSV

df_Season_Advance = pd.read_csv('e:\\nba\\Season_Advance_Totals.csv', header=0)

df_Season_Advance.info()
#df_Season_Advance.describe()

### Season EFF Calculation

def EFF_Calculation(x):
    PTS = x[29]
    TRB = x[23]
    AST = x[24]
    STL = x[25]
    BLK = x[26]
    FGA = x[9]
    FG  = x[8] 
    FTA = x[19]
    FT  = x[18]
    TOV = x[27]
    G   = x[5]
    
    EFF = ((PTS + TRB + AST + STL + BLK) -(FGA-FG)-(FTA-FT)-TOV) / G
    return EFF


df_Season_Total['EFF'] = df_Season_Total.apply(EFF_Calculation,axis =1)
df_Season_Total[['Season','Name','EFF']].head(10)

df_Season_Total.info()

### Season GmSc Calculation

def GmSc_Calculation(x):
    PTS = x[29]
    ORB = x[21]
    DRB = x[22]    
    AST = x[24]
    STL = x[25]
    BLK = x[26]
    FGA = x[9]
    FG  = x[8] 
    FTA = x[19]
    FT  = x[18]
    PF  = x[28]
    TOV = x[27]
    G   = x[5]
    
    GmSc = (PTS + 0.4 * FG - 0.7 * FGA - 0.4*(FTA - FT)
            + 0.7 * ORB + 0.3 * DRB + STL + 0.7 * AST + 0.7 * BLK - 0.4 * PF - TOV) / G
    return GmSc


df_Season_Total['GmSc'] = df_Season_Total.apply(GmSc_Calculation,axis =1)
df_Season_Total[['Season','Name','EFF','GmSc']].head(10)

###Merge Season_Total Field ['ORtg','DRtg','EFF','GmSc'] to Season_Advance

df_Season_Advance = pd.merge(df_Season_Advance, df_Season_Total[['Season', 'Tm','Name','ORtg','DRtg','EFF','GmSc']], on=['Season', 'Tm','Name'])
print df_Season_Advance.sort_index(by=['Season','Tm','Name']).head(10)

###Read Player Playoffs CSV

df_playoffs_Total = pd.read_csv('e:\\nba\\Playoffs_Totals.csv', header=0)

df_playoffs_Total.info()

###Read Playoffs_Advance_Totals CSV

df_playoffs_Advance = pd.read_csv('e:\\nba\\Playoffs_Advance_Totals.csv', header=0)

df_playoffs_Advance.info()

###Playoffs EFF Calculation

df_playoffs_Total['EFF'] = df_playoffs_Total.apply(EFF_Calculation,axis =1)
df_playoffs_Total[['Season','Name','EFF']].head(10)

###Playoffs GmSc Calculation

df_playoffs_Total['GmSc'] = df_playoffs_Total.apply(GmSc_Calculation,axis =1)
df_playoffs_Total[['Season','Name','EFF','GmSc']].head(10)

###Merge Playoffs_Total Field ['ORtg','DRtg','EFF','GmSc'] to Playoffs_Advance

df_playoffs_Advance = pd.merge(df_playoffs_Advance, df_playoffs_Total[['Season', 'Tm','Name','ORtg','DRtg','EFF','GmSc']], on=['Season', 'Tm','Name'])
print df_playoffs_Advance.sort_index(by=['Season','Tm','Name']).head(10)

###Convert Season_Advance column [Salary] To Integer

def Salary_Convert(x):
    Salary = x[29] 
    # Salary : "$2,126,520"
    Salary_Int = Salary.replace('"', "").replace(',', "").replace('$', "")    
    return Salary_Int

#df_Season_Advance.info()
df_Season_Advance['Salary'] = df_Season_Advance.apply(Salary_Convert,axis =1)
print df_Season_Advance.sort_index(by=['Season','Tm','Name']).head(10)

###Convert Playoffs_Advance column [Salary] To Integer

df_playoffs_Advance['Salary'] = df_playoffs_Advance.apply(Salary_Convert,axis =1)
print df_playoffs_Advance.sort_index(by=['Season','Tm','Name']).head(10)

###Output DataFrame to CSV ( with Name Column)

def Output_CSV(output_df,location):
    output_file = open(location, "ab")
    output_df.to_csv(output_file,index=False)
    output_file.close()
    
Season_Summary_location = "e:\\nba\\Season_Collection_with_Name.csv"
Output_CSV(df_Season_Advance.sort_index(by=['Season','Tm']),Season_Summary_location)

Playoffs_Summary_location = "e:\\nba\\Playoffs_Collection_with_Name.csv"
Output_CSV(df_playoffs_Advance.sort_index(by=['Season','Tm']),Playoffs_Summary_location)

###Remove column [Name]

df_Season_Advance = df_Season_Advance.drop('Name', 1)
df_playoffs_Advance = df_playoffs_Advance.drop('Name', 1)

###Missing value

df_Season_Advance = df_Season_Advance.fillna(0)
df_playoffs_Advance = df_playoffs_Advance.fillna(0)

###Output DataFrame to CSV

def Output_CSV(output_df,location):
    output_file = open(location, "ab")
    output_df.to_csv(output_file,index=False)
    output_file.close()
    
Season_Summary_location = "e:\\nba\\Season_Collection.csv"
Output_CSV(df_Season_Advance.sort_index(by=['Season','Tm']),Season_Summary_location)

Playoffs_Summary_location = "e:\\nba\\Playoffs_Collection.csv"
Output_CSV(df_playoffs_Advance.sort_index(by=['Season','Tm']),Playoffs_Summary_location)
