==> 5.Basketball-reference Data Normalization and Regression.ipynb <==
#Basketball-reference Data Normalization and Regression

# special IPython command to prepare the notebook for matplotlib
from collections import defaultdict
from pandas.stats.api import ols
from sklearn.linear_model import LinearRegression

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline
import re
from sklearn import preprocessing
#import statsmodels.api as sm
#import statsmodels as sm


### Read Player Seaaon CSV

df_Season_Collect = pd.read_csv('e:\\nba\\Season_Collection.csv', header=0)

df_Season_Collect.info()
#df_Season_Collect.describe()

year_list = df_Season_Collect['Season'].unique()
print year_list
yearStr = year_list[2]

year = df_Season_Collect['Season'].map(lambda x: x == yearStr)
df_year = df_Season_Collect[year]
df_year

###Normalization Function

def Normalization(df_data):
    df_norm = ((df_data - df_data.min()) / (df_data.max() - df_data.min()))
    return df_norm

###Column Index

norm_column = ['Age','G','MP','EFF','GmSc','PER','ORtg','DRtg','AST_Per','STL_Per','BLK_Per',
             'TOV_Per','TS_Per','FT_Per','three_Per','ORB_Per','DRB_Per','TRB_Per','USG_Per','Salary']

###Season Data Normalization

df_tmp = df_year[norm_column] 
df_norm = Normalization(df_tmp)
df_norm

df_Season_norm = pd.DataFrame(columns=df_norm.columns)
df_Season_norm

for yearStr in year_list:
    year = df_Season_Collect['Season'].map(lambda x: x == yearStr)
    df_year = df_Season_Collect[year]
    df_data = df_year[norm_column]
    df_norm = Normalization(df_data)    
    df_Season_norm = df_Season_norm.append(df_norm)

df_Season_norm.info()
df_Season_norm.head(10)

###Statistical indicators

vars_list = ['Age','G','MP','EFF','GmSc','PER','ORtg','DRtg','AST_Per','STL_Per','BLK_Per',
             'TOV_Per','TS_Per','FT_Per','three_Per','ORB_Per','DRB_Per','TRB_Per','USG_Per']

###Regression

res_Season = ols(y=df_Season_Collect['Salary'], x=df_Season_Collect[vars_list])
res_Season

###Normalization Regression

res_Season_norm = ols(y=df_Season_norm['Salary'], x=df_Season_norm[vars_list])
res_Season_norm

###Output Season Normalization Data TO CSV 

def Output_CSV(output_df,location):
    output_file = open(location, "ab")
    output_df.to_csv(output_file,index=False)
    output_file.close()
    
Season_Normalization_location = "e:\\nba\\Season_Normalization.csv"
Output_CSV(df_Season_norm,Season_Normalization_location)

###Read Player Playoffs CSV

df_Playoffs_Collect = pd.read_csv('e:\\nba\\Playoffs_Collection.csv', header=0)

df_Playoffs_Collect.info()

###Playoffs Data Normalization 

df_Playoffs_norm = pd.DataFrame(columns=df_norm.columns)
df_Playoffs_norm

for yearStr in year_list:
    year = df_Playoffs_Collect['Season'].map(lambda x: x == yearStr)
    df_year = df_Playoffs_Collect[year]
    df_data = df_year[norm_column]
    df_norm = Normalization(df_data)    
    df_Playoffs_norm = df_Playoffs_norm.append(df_norm)

df_Playoffs_norm.info()
df_Playoffs_norm.head(10)

###Regression Playoffs

res_playoffs = ols(y=df_Playoffs_Collect['Salary'], x=df_Playoffs_Collect[vars_list])
res_playoffs

###Playoffs Normalization Regression

res_Playoffs_norm = ols(y=df_Playoffs_norm['Salary'], x=df_Playoffs_norm[vars_list])
res_Playoffs_norm

###Output Playoffs Normalization Data TO CSV¶

Playoffs_Normalization_location = "e:\\nba\\Playoffs_Normalization.csv"
Output_CSV(df_Playoffs_norm,Playoffs_Normalization_location)
