==> 6.1 Basketball-reference Quartile Calc with Name.ipynb <==
#Basketball-reference Quartile Calc

# special IPython command to prepare the notebook for matplotlib
from collections import defaultdict

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline
import re



### Read Player Seaaon CSV

df_Season_Collect = pd.read_csv('e:\\nba\\Season_Collection_with_Name.csv', header=0)

df_Season_Collect.info()
#df_Season_Collect.describe()

year_list = df_Season_Collect['Season'].unique()
print year_list
yearStr = year_list[1]

def getYearQuartile(df_year):
    df_year['Salary'].quantile([0,.25,.5,.75,1])
    Quartile_list = (Min,Q1,Q2,Q3,Q4) =  df_year['Salary'].quantile([0,.25,.5,.75,1])  
    df_year.loc[:,'Q1'] = Q1
    df_year.loc[:,'Q2'] = Q2
    df_year.loc[:,'Q3'] = Q3
    df_year.loc[:,'Q4'] = Q4   
    
    return df_year

def CheckQuartile(x):
    Q1 = x[34]
    Q2 = x[35]
    Q3 = x[36]
    Q4 = x[37]
    Salary = x[29]
    
    if (Salary <= Q1):
        result = "Q1"
    elif (Salary > Q1) & (Salary <= Q2):
        result = "Q2"
    elif (Salary > Q2) & (Salary <= Q3):
        result = "Q3"
    elif (Salary > Q3) & (Salary <= Q4):
        result = "Q4"
    else:
        result = "Q2"
    
    return result

year = df_Season_Collect['Season'].map(lambda x: x == yearStr)
df_year = df_Season_Collect[year]
#df_year
df_year.info()

df_result = getYearQuartile(df_year)
df_result.loc[:,'Quartile'] = df_result.apply(CheckQuartile,axis =1)
df_result.info()
df_result

#df_output = pd.DataFrame(data=None, columns=df_result.columns,index=0)
name_list = list(df_result.columns)
df_Season_output = pd.DataFrame(columns=name_list)
df_Season_output

def CalcQuartile(df_data):
    df_output = pd.DataFrame(columns=name_list)
    for yearStr in year_list:
        year = df_data['Season'].map(lambda x: x == yearStr)
        df_year = df_data[year]
        df_result = getYearQuartile(df_year)
        df_result.loc[:,'Quartile'] = df_result.apply(CheckQuartile,axis =1)
        df_output = df_output.append(df_result)
    
    return df_output

df_Season_output = CalcQuartile(df_Season_Collect)
df_Season_output.info()

###Read Player Playoffs CSV

df_Playoffs_Collect = pd.read_csv('e:\\nba\\Playoffs_Collection_with_Name.csv', header=0)

df_Playoffs_Collect.info()
#df_Playoffs_Collect.describe()

year_list = df_Playoffs_Collect['Season'].unique()
print year_list

name_list = list(df_result.columns)
df_Playoffs_output = pd.DataFrame(columns=name_list)
df_Playoffs_output

df_Playoffs_output = CalcQuartile(df_Playoffs_Collect)
df_Playoffs_output.info()

###Output DataFrame to CSV

def Output_CSV(output_df,location):    
    output_file = open(location, "ab")
    output_df.to_csv(output_file,index=False)
    output_file.close()

Season_location = "e:\\nba\\Season_Quartile_with_Name.csv"
Playoffs_location = "e:\\nba\\Playoffs_Quartile_with_Name.csv"

Output_CSV(df_Season_output,Season_location)
Output_CSV(df_Playoffs_output,Playoffs_location)
