==> 8.Basketball-reference Data Calc Compare LastYearEFF.ipynb <==
#Basketball-reference Data GroupBy 

# special IPython command to prepare the notebook for matplotlib
from collections import defaultdict

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline
import re



### Read Player Season CSV

df_Season = pd.read_csv('e:\\nba\\Season_Quartile_with_Name.csv', header=0)

df_Season.info()
#df_Season.describe()

###GroupBy

Season_grouped = df_Season.groupby('Name')
len(Season_grouped)

def compare(a,b):
    if a > b:
        return 1
    elif a< b:
        return -1
    else:
        return 0

def EFFcalc(length,df_tmp):
    df_result = pd.DataFrame(columns=name_list)
    #print "length",len(df_tmp)
    
    for i in xrange(length):
        if i == 0:
            curt = df_tmp.iloc[i].values.tolist()            
            curt.append(0)            
            df_result.loc[len(df_result)] = curt
            #print len(df_result),df_result[['Season','Name']]
            #print "length",len(df_result)
        else:
            curt = df_tmp.iloc[i].values.tolist()
            last = df_tmp.iloc[i-1].values.tolist()
            result = compare(curt[32],last[32])            
            
            curt.append(result)
            #print "curt in body",curt
            df_result.loc[len(df_result)] = curt
        
    #print df_result[['Season','Name']]
    #print "length",len(df_result)
    return df_result

def GroupDataCalc (df_grouped):
    df_output = pd.DataFrame(columns=name_list)    
    for Name,df_tmp in df_grouped:
        if len(df_tmp) == 1:            
            curt = df_tmp.iloc[0].values.tolist()
            curt.append(0)        
            df_output.loc[len(df_output)] = curt
            #df_Season_output.append(df_tmp)
        else:            
            df_result = EFFcalc(len(df_tmp),df_tmp)
            df_output = df_output.append(df_result)
    return df_output


name_list = list(df_Season.columns)
name_list.append('LastYearEFF')

df_Season_result = GroupDataCalc(Season_grouped)
df_Season_result.info()
#df_Season_result.head(20)

###Read Player Playoffs CSV

df_Playoffs = pd.read_csv('e:\\nba\\Playoffs_Quartile_with_Name.csv', header=0)

df_Playoffs.info()
#df_Playoffs.describe()

Playoffs_grouped = df_Playoffs.groupby('Name')
len(Playoffs_grouped)

df_Playoffs_result = GroupDataCalc(Playoffs_grouped)
df_Playoffs_result.info()
#f_Playoffs_result.head(20)

###Feature Column to Weka

Feature_list = ['Age','G','MP','EFF','GmSc','PER','ORtg','DRtg','AST_Per','STL_Per','BLK_Per','TOV_Per','TS_Per',
                'FT_Per','three_Per','ORB_Per','DRB_Per','TRB_Per','USG_Per','LastYearEFF','Quartile']

#df_Season_output = df_Season_result[Feature_list]
df_Season_output.fillna(0)
df_Season_output.info()
df_Season_output.head(5)

#df_Playoffs_output = df_Playoffs_result[Feature_list]
df_Playoffs_output.fillna(0)
df_Playoffs_output.info()
df_Playoffs_output.head(5)

###Output DataFrame to CSV

def Output_CSV(output_df,location):    
    output_file = open(location, "ab")
    output_df.to_csv(output_file,index=False)
    output_file.close()

Season_location = "e:\\nba\\Season_Weka.csv"
Playoffs_location = "e:\\nba\\Playoffs_Weka.csv"

Output_CSV(df_Season_output,Season_location)
Output_CSV(df_Playoffs_output,Playoffs_location)
